refactor(userspace): extract tunnel wire types to protocol_tunnels.go
Fourth code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the tunnel-endpoint config snapshots and the WireGuard
peer/tunnel runtime status types into a cohesive sibling
protocol_tunnels.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change):
TunnelEndpointSnapshot, TunnelWgPeerWire, WgPeerStatus, WgTunnelStatus.

WgPeerStatus/WgTunnelStatus mirror the Rust types in
userspace-dp/src/protocol/control.rs; their json tags must stay
identical on both sides and are preserved byte-identical here. Build
passes; gofmt clean.
